@@ -9,8 +9,8 @@
     "cors": "^2.8.5",
     "express": "^4.16.4",
     "moxios": "^0.4.0",
-    "react": "^16.6.1",
-    "react-dom": "^16.6.1",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
     "react-redux": "^5.1.1",
     "react-scripts": "^2.1.3",
     "redux": "^4.0.1",
